+++ src/manips/SoDirectionalLightManip.cpp
@@ -279,7 +279,7 @@ SoDirectionalLightManip::getBoundingBox(SoGetBoundingBoxAction * action)
     action->resetCenter();
   }
   if (numcenters != 0) {
-    action->setCenter(center / numcenters, FALSE);
+    action->setCenter(center / (float) numcenters, FALSE);
   }
 }
