+++ src/manips/SoPointLightManip.cpp
@@ -280,7 +280,7 @@ SoPointLightManip::getBoundingBox(SoGetBoundingBoxAction * action)
     action->resetCenter();
   }
   if (numcenters != 0) {
-    action->setCenter(center / numcenters, FALSE);
+    action->setCenter(center / (float) numcenters, FALSE);
   }
 }
